+++ apps/scan/src/main/scala/org/lfdecentralizedtrust/splice/scan/admin/api/client/commands/HttpScanAppClient.scala
@@ -1502,7 +1502,7 @@ object HttpScanAppClient {
               new allocationinstructionv1.AllocationFactory_Allocate(
                 choiceArgs.expectedAdmin,
                 choiceArgs.allocation,
-                choiceArgs.inputHoldings,
+                choiceArgs.inputHoldingCids,
                 new metadatav1.ExtraArgs(
                   choiceContext,
                   choiceArgs.extraArgs.meta,